@@ -11,10 +11,10 @@
 	},
 	"devDependencies": {
 		"@sveltejs/kit": "^2.17.2",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/vite-plugin-svelte": "^4.0.4",
 		"autoprefixer": "^10.4.20",
 		"postcss": "^8.5.2",
-		"svelte": "^4.2.19",
+		"svelte": "^5.20.1",
 		"svelte-check": "^4.1.4",
 		"tailwindcss": "^3.4.17",
 		"typescript": "^5.7.3",
